% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_queries.R
\name{ADULTS}
\alias{ADULTS}
\title{ADULTS}
\usage{
ADULTS(username = "bt", host = "scidb.mpio.orn.mpg.de")
}
\arguments{
\item{username}{user}

\item{host}{host}
}
\description{
ADULTS
}
\examples{
require(dygraphs)
x = ADULTS()[, week := week(capture_date)]
dygraph(x)

}
