% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AAA.R
\docType{package}
\name{bib2}
\alias{bib2}
\alias{.onLoad}
\alias{bib2-package}
\title{Breding BIrd Planer (v2)}
\usage{
The user interface is installed at in at http://scidb.mpio.orn.mpg.de
}
\description{
Mapping and field work organiser
}
