% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_elements.R
\name{map_empty}
\alias{map_empty}
\title{Maps}
\usage{
map_empty()
}
\description{
Maps
}
\examples{
map_base()
x = copy(boxesxy)

x[, ':=' (g = cut(1:277, 4), f = ifelse( as.numeric(box) < 100, 's', 'l') ) ]
map_base(x, x = 'long', y = 'lat', label = 'box', size = 2, color = 'g')
map_base(x, x = 'long', y = 'lat', label = 'box', size = 2, color = 'f')

map_base(x, x = 'long', y = 'lat', label = 'box', size = 2, family = 'FreeSansBoldOblique')

}
