% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_elements.R
\name{map_geom_label}
\alias{map_geom_label}
\title{Map w. geom_label()}
\usage{
map_geom_label(dat = boxesxy, size = 2.5, ...)
}
\description{
Map w. geom_label()
}
\examples{
map_geom_label(x = 'long', y = 'lat', label = 'box3d', size = 2)
x = copy(boxesxy)

x[, ':=' (g = cut(1:277, 4), f = ifelse( as.numeric(box) < 100, 's', 'l') ) ]
map_geom_label(x = 'long', y = 'lat', label = 'box3d', size = 2, color = 'g', dat = x)
map_geom_label(x = 'long', y = 'lat', label = 'box3d', size = 2, color = 'f', fill = 'g', dat = x)

}

