% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_elements.R
\name{maps}
\alias{maps}
\alias{map_empty}
\alias{map_base}
\alias{map_nests}
\title{maps}
\usage{
map_empty()

map_base(size = 2.5, family = "Times", fontface = "bold",
  bottom_ann = paste("Printed on", format(Sys.time(),
  "\%a, \%d \%b \%y \%H:\%M")))

map_nests(n, size = 2.5, family = "Times", fontface = "bold",
  title = paste("made on:", Sys.Date()), bottom_ann = paste("Printed on",
  format(Sys.time(), "\%a, \%d \%b \%y \%H:\%M")))
}
\arguments{
\item{n}{a data.table returned by  nests()}

\item{title}{goes to ggtitle (should be the reference date)}
}
\examples{
map_base(family = 'Times')
 rdate = Sys.Date() - 1
 n = nests(rdate) \%>\% nest_state(ref_date  =rdate )
 map_nests(n)
}
