% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_elements.R
\name{nest_legend}
\alias{nest_legend}
\title{Legend
Nest stages, frequencies and late/early score}
\usage{
nest_legend(n)
}
\arguments{
\item{n}{a  data.table (commented with ref date) returned by nests()}

\item{N}{a  data.table returned by allNests()}
}
\description{
Legend
Nest stages, frequencies and late/early score
}
\examples{
n = nests(Sys.Date() ) \%>\% nest_state()
nest_legend(n)
}
