% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_queries.R
\name{nest_state}
\alias{nest_state}
\title{nest_state}
\usage{
nest_state(x)
}
\arguments{
\item{x}{a commented [using setattr] (with ref date) data.table returned by nests()}

\item{ref_date}{reference date}
}
\description{
nest_state
}
\examples{
nests() \%>\% nest_state()
nests("2016-04-12") \%>\% nest_state()
}
