% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_elements.R
\name{nest_table}
\alias{nest_table}
\title{Nests table}
\usage{
nest_table(n, file = tempfile(fileext = ".xlsx"))
}
\arguments{
\item{n}{a data.table returned by  nests()}
}
\value{
path to output file
}
\description{
Nests table for printing
}
\examples{
\dontrun{
 x = nests(Sys.Date() - 1 )
 n = nest_state(x, hatchingModel = predict_hatchday_model(Breeding(), rlm) )
 p = nest_table(n)
}

}
