% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_queries.R
\name{phenology}
\alias{phenology}
\title{nest_state}
\usage{
phenology(minimum_stage = "C")
}
\arguments{
\item{minimum_stage}{(one of "LT", "B", "C", "LIN", "firstEgg","hatchDate","fledgeDate") default to "C"}

\item{x}{a data.table as returned by nests()}

\item{nesr_stages}{character vector}

\item{hatchingModel}{a model to use for prediction
phenology}
}
\description{
nest_state
}
\examples{
phenology()
}
