% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_elements.R
\name{plots}
\alias{plots}
\alias{plot_phenology_firstDates}
\title{plots}
\usage{
plot_phenology_firstDates(refdate = Sys.Date(),
  x = phenology_firstDates(refdate))
}
\examples{
  all_phenology = phenology()
plot_phenology_firstDates()
}
