% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predictions.R
\name{predict_firstEgg}
\alias{predict_firstEgg}
\alias{predict_firstEgg_data}
\title{predict_firstEgg}
\usage{
predict_firstEgg(dat, v, refyear = year(Sys.Date()))

predict_firstEgg_data(F, refdate = Sys.Date())
}
\description{
predict_firstEgg
}
\examples{
date_ = Sys.Date()
F = phenology('LIN')
predFirstEggData = predict_firstEgg_data(F, date_ )
predFirstEgg = predict_firstEgg(predFirstEggData, yday(date_), year(date_) )[, date_ := as.POSIXct(date_) ]

ggplot(predFirstEggData, aes(y = first_Egg , x = first_Lining, x) ) + 
 geom_point() + 
 geom_smooth(method = 'lm') + geom_text(aes(label = year_), vjust= 'bottom') + 
 geom_pointrange(data = predFirstEgg, aes(x = date_, y = fit, ymin = lwr, ymax = upr ), col = 2) + 
 ggtitle(paste(predFirstEgg[, difftime(fit, date_, units = 'days')\%>\% round], 'days till first egg.'))
}
