% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predictions.R
\name{predict_hatchday_model}
\alias{predict_hatchday_model}
\title{predict_hatchday_model}
\usage{
predict_hatchday_model(x, model = lm)
}
\description{
predict_hatchday_model
}
\examples{
\donttest{
predict_hatchday_model( Breeding() )
}
}
