% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_queries.R
\name{RINGS_FOR_RADO}
\alias{RINGS_FOR_RADO}
\title{RINGS_FOR_RADO}
\usage{
RINGS_FOR_RADO(
  seasons = c(2018),
  new_ring = data.table(ID = c("B4H1627"), date_time = c("2014-11-08 10:02:00")),
  not_banded_by_us = c("TN90629")
)
}
\arguments{
\item{seasons}{season(s) from which the data is required. Can be vector or number. 
Note that this refers to the season, not to the calendar year 
(e.g. 2018 is summer 2017 through spring 2018)}

\item{new_ring}{date and time when a bird recieved a new metal band. Only necessary if 
there is any chance that the bird was caught in between somewhere else}

\item{not_banded_by_us}{band numbers of birds that were banded elsewhere should be provided as a vector}
}
\description{
RINGS_FOR_RADO
}
\examples{
path = RINGS_FOR_RADO(seasons = 2018)
}
