% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_elements.R
\name{map_empty}
\alias{map_base}
\alias{map_empty}
\title{Maps}
\usage{
map_empty()

map_base(dat = boxesxy, size = 2.5, family = "Times", fontface = "bold",
  x = "long", y = "lat", label = "box", bottom_ann = paste("Printed on",
  format(Sys.time(), "\%a, \%d \%b \%y \%H:\%M")), ...)
}
\description{
Maps
}
\examples{
map_base()
map_base(family = 'Arial')

x = copy(boxesxy)
x[, ':=' (g = cut(1:277, 4), f = ifelse( as.numeric(box) < 100, 's', 'l') ) ]

map_base(x,  color = 'g')
map_base(x, label = 'box3d', size = 2, color = 'f', fill = 'g')



}

