% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_elements.R
\name{maps}
\alias{maps}
\alias{theme_bib2}
\alias{map_legend}
\alias{print_ann}
\alias{map_empty}
\alias{map_base}
\alias{map_nests}
\title{maps}
\usage{
theme_bib2()

map_legend(size = 2.5, right = "box", left = "checked days ago",
  top = "stage age/days till hatch/chick age",
  bottom = "eggs|chicks(?=guessed)", x = 4417700, y = 5334960)

print_ann(color = "grey", x = Inf, y = Inf, vjust = "bottom",
  hjust = "top", angle = 90)

map_empty()

map_base(size = 2.5, family = "Times", fontface = "bold",
  printdt = FALSE)

map_nests(n, size = 2.5, family = "Times", fontface = "bold",
  title = paste("made on:", Sys.Date()), printdt = FALSE)
}
\arguments{
\item{n}{a data.table returned by  nests()}

\item{title}{goes to ggtitle (should be the reference date)}
}
\value{
a list of geoms defining the legend around the box
}
\examples{
map_empty() 
map_base(family = 'Times') 
 rdate = Sys.Date() - 1
 n = nests(rdate) \%>\% nest_state( )
 map_nests(n)  
}
