% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_elements.R
\name{maps}
\alias{maps}
\alias{theme_bib2}
\alias{map_legend}
\alias{print_ann}
\alias{map_empty}
\alias{map_base}
\alias{map_nests}
\alias{map_experiment}
\title{maps}
\usage{
theme_bib2()

map_legend(
  size = 2.5,
  right = "box",
  left = "checked days ago",
  top = "stage age|days to hatch",
  bottom = "eggs|chicks",
  x = 543,
  y = 735
)

print_ann(
  color = "grey",
  x = Inf,
  y = Inf,
  vjust = "bottom",
  hjust = "top",
  angle = 90
)

map_empty()

map_base(size = 2.5, family = "sans", fontface = "plain", printdt = FALSE)

map_nests(
  n,
  size = 2.5,
  family = "sans",
  fontface = "plain",
  title = paste("made on:", Sys.Date()),
  notes = "",
  nx = -20,
  ny = 650
)

map_experiment(exp_id)
}
\arguments{
\item{n}{a data.table returned by  nests()}

\item{title}{goes to ggtitle (should be the reference date)}

\item{notes}{notes under legend annotations}

\item{nx}{notes x location}

\item{ny}{notes y location}

\item{exp_id}{the id of the experiment as defined in the experiments table.}
}
\value{
a list of geoms defining the legend around the box

a list of geoms to append to map_nests()
}
\description{
maps
}
\examples{
map_empty() 
map_base(family = 'sans') 
 x = nests(Sys.Date() - 1 )
 notes = c('note 1: this is note 1\nnote 99: this is note 99\nnote 9999+1: this is note 9999+1')
 n = nest_state(x, hatchingModel = predict_hatchday_model(Breeding(), rlm) )
 map_nests(n)  
 map_nests(n, notes = notes) + print_ann() 


 x = map_experiment()

}
