% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_elements.R
\name{nest_state}
\alias{nest_state}
\title{nest_state}
\usage{
nest_state(x, nest_stages = NULL, hatchingModel)
}
\arguments{
\item{x}{a data.table as returned by nests()}

\item{hatchingModel}{a model to use for prediction}

\item{nesr_stages}{character vector}
}
\description{
nest_state
}
\examples{
 B = Breeding()
hatchingModel = predict_hatchday_model(B, MASS::rlm)  
nests() \%>\% nest_state()
nests("2016-04-12") \%>\% nest_state()
}
